Move scanner API test to integration tests
